feat(WeaponStats): Added small randomness to stats outcomes

+++ Endless_Survivor/Assets/Scripts/Weapons/Weapon.cs
@@ -39,7 +39,7 @@ public class Weapon : MonoBehaviour
             return;
 
         _canShoot = false;
-        _shootCooldown = 1f / (float)(WeaponStats.AttackSpeed + GameManager.gm.selectedCharacter.PlayerStats.AttackSpeed);
+        _shootCooldown = 1f / WeaponStats.AttackSpeed + GameManager.gm.selectedCharacter.PlayerStats.AttackSpeed;
         Attack();
     }
